Improve the user list page.

+++ com.jotunheim.mimir.domain/src/main/java/com/jotunheim/mimir/domain/data/RoleAccessLevel.java
@@ -12,7 +12,7 @@ public class RoleAccessLevel {
     public static final int ADMIN = 8;
     public static final int SUPERVISOR = 16;
 
-    public static final int  ROLE_ID_USER = 1;
-    public static final int  ROLE_ID_ADMIN = 2;
-    public static final int  ROLE_ID_SUPERVISOR = 3;
+    public static final long  ROLE_ID_USER = 1;
+    public static final long  ROLE_ID_ADMIN = 2;
+    public static final long  ROLE_ID_SUPERVISOR = 3;
 }
